moved queries and transactions out of contract.rs
