+++ contracts/vesting/src/lib.rs
@@ -6,8 +6,10 @@
 
 pub mod contract;
 mod errors;
+mod queries;
 mod queued_migrations;
 mod storage;
 mod support;
 mod traits;
+mod transactions;
 pub mod vesting;
